refactor(DetailedView) added venue details component for code readability
